// src/app/models/event.ts
// Generated by https://quicktype.io

export interface EventModel {
    eventId:     number;
    name:        string;
    description: string;
    imageId:     number;
    startDate:   string;
    endDate:     string;
    expired:     boolean;
    active:      boolean;
    image:       Image;
    city:        string;
    state:       string;
    users:       User[];
}

export interface Image {
    imageId: number;
    path:    string;
}

export interface User {
    userId:    number;
    firstName: string;
    lastName:  string;
    userName:  string;
    bio:       string;
    image:     Image;
}